Rename packages for consistency + fix warning at startup (mysql ssl disabled)
