--- infrastructure/mysql-persistence/src/main/java/com/rdelgatte/hexagonal/mysqlpersistence/MysqlPersistenceApplication.java
@@ -1,12 +1,0 @@
-package com.rdelgatte.hexagonal.mysqlpersistence;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MysqlPersistenceApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(MysqlPersistenceApplication.class, args);
-  }
-}
