--- infrastructure/mysql-persistence/src/main/java/com/rdelgatte/hexagonal/mysqlpersistence/repository/ProductRepository.java
@@ -1,11 +1,0 @@
-package com.rdelgatte.hexagonal.mysqlpersistence.repository;
-
-import com.rdelgatte.hexagonal.mysqlpersistence.model.MysqlProduct;
-import io.vavr.control.Option;
-import java.util.UUID;
-import org.springframework.data.repository.CrudRepository;
-
-interface ProductRepository extends CrudRepository<MysqlProduct, UUID> {
-
-  Option<MysqlProduct> findOneByCode(String code);
-}
