+++ infrastructure/mysql-persistence/src/main/java/com/rdelgatte/hexagonal/persistence/mysql/repository/ProductRepository.java
@@ -1,0 +1,11 @@
+package com.rdelgatte.hexagonal.persistence.mysql.repository;
+
+import com.rdelgatte.hexagonal.persistence.mysql.model.MysqlProduct;
+import io.vavr.control.Option;
+import java.util.UUID;
+import org.springframework.data.repository.CrudRepository;
+
+interface ProductRepository extends CrudRepository<MysqlProduct, UUID> {
+
+  Option<MysqlProduct> findOneByCode(String code);
+}
