+++ infrastructure/mysql-persistence/src/main/resources/application.yml
@@ -1,8 +1,8 @@
 spring:
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
   datasource:
-    url: jdbc:mysql://localhost:3306/hexagonal
+    url: jdbc:mysql://localhost:3306/hexagonal?useSSL=false
     username: hexagonal
     password: hexagonal
